Added arena initialization calls.
The arena will need to be made dynamic at some point in the future,
because currently it stops at 50 members, but usage may easily involve
many more members.

It may also be necessary to assess probable use cases, as arena
allocation and deallocation might not make sense outside of a testing
context.